requirements.txt
langchain>=0.1.0
langgraph>=0.0.11
llama-cpp-python>=0.2.11
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
fastapi>=0.109.0
uvicorn>=0.27.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pydantic>=2.5.0
